Description: This pull request introduces the Node and LinkedList classes. The Node class represents individual nodes in the linked list, while the LinkedList class provides functionality for adding nodes and printing the list.
Implementation Details:
Node.java: A class representing a node containing an integer and a
reference to the next node.
LinkedList.java: Implements a linked list that allows adding nodes to
the end and printing the list.
Main.java: Provides a simple interface for users to add integers to the
list and print the list contents.
Conclusion:
The linked list implementation is functional. Users can input integers
to dynamically build a list, and the program will display the list once
the user decides to exiting by typing '5'